// models/_project.ts
export const projectModel = {
  languages: {
    default: {
      codename: 'default',
      name: 'Default project language',
    },
  },
  contentTypes: {
    _component__cta: {
      codename: '_component__cta',
      name: '[Component] CTA',
      elements: {
        button_link: {
          codename: 'button_link',
          name: 'Button link',
        },
        button_text: {
          codename: 'button_text',
          name: 'Button text',
        },
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        subheadline: {
          codename: 'subheadline',
          name: 'Subheadline',
        },
      },
    },
    _component__hero: {
      codename: '_component__hero',
      name: '[Component] Hero',
      elements: {
        image: {
          codename: 'image',
          name: 'Image',
        },
        subheadline: {
          codename: 'subheadline',
          name: 'Subheadline',
        },
        cta_text: {
          codename: 'cta_text',
          name: 'CTA text',
        },
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        cta_link: {
          codename: 'cta_link',
          name: 'CTA link',
        },
      },
    },
    _component__text_with_image: {
      codename: '_component__text_with_image',
      name: '[Component] Text with image',
      elements: {
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        text: {
          codename: 'text',
          name: 'Text',
        },
        image: {
          codename: 'image',
          name: 'Image',
        },
      },
    },
    article_example_content_type: {
      codename: 'article_example_content_type',
      name: 'Article—example content type',
      elements: {
        body: {
          codename: 'body',
          name: 'Body',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        url: {
          codename: 'url',
          name: 'URL',
        },
      },
    },
    landing_page: {
      codename: 'landing_page',
      name: 'Landing page',
      elements: {
        content: {
          codename: 'content',
          name: 'Content',
        },
        url_slug: {
          codename: 'url_slug',
          name: 'URL slug',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
      },
    },
    landing_page_example_content_type: {
      codename: 'landing_page_example_content_type',
      name: 'Landing page—example content type',
      elements: {
        body: {
          codename: 'body',
          name: 'Body',
        },
        product_list: {
          codename: 'product_list',
          name: 'Product list',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        url: {
          codename: 'url',
          name: 'URL',
        },
      },
    },
    product_example_content_type: {
      codename: 'product_example_content_type',
      name: 'Product—example content type',
      elements: {
        url: {
          codename: 'url',
          name: 'URL',
        },
        description: {
          codename: 'description',
          name: 'Description',
        },
        name: {
          codename: 'name',
          name: 'Name',
        },
        image: {
          codename: 'image',
          name: 'Image',
        },
        taxonomy___categorizing_target_audience: {
          codename: 'taxonomy___categorizing_target_audience',
          name: 'Taxonomy - categorizing target audience',
        },
      },
    },
  },
  taxonomies: {
    taxonomy_categorizing_target_audience: {
      codename: 'taxonomy_categorizing_target_audience',
      name: 'Taxonomy—categorizing target audience',
      terms: {
        for_other_businesses__b2b_: {
          codename: 'for_other_businesses__b2b_',
          name: 'For other businesses (B2B)',
          terms: {},
        },
        for_consumers__b2c_: {
          codename: 'for_consumers__b2c_',
          name: 'For consumers (B2C)',
          terms: {},
        },
      },
    },
  },
};